Add bin scripts and move code to src dir
