+++ bin/modify-composer
@@ -1,0 +1,15 @@
+#!/usr/bin/env node
+
+const modifyComposer = require('../src/modifyComposer');
+const showHelp = require('../src/showHelp');
+
+if (process.argv.length < 5) {
+  showHelp();
+  process.exit(1);
+}
+
+const authToken = process.argv[2];
+const jobName = process.argv[3];
+const isPackage = (process.argv[4] === 'package');
+
+modifyComposer(authToken, jobName, isPackage);
