+++ src/getDependencies.js
